+++ js/apps/system/_admin/aardvark/APP/react/src/views/collections/indices/addIndex/invertedIndex/InvertedIndexFieldsDataEntry.tsx
@@ -1,7 +1,7 @@
 import { Box } from "@chakra-ui/react";
 import { useFormikContext } from "formik";
 import React from "react";
-import { FormField, FormFieldProps } from "../FormField";
+import { FormField, FormFieldProps } from "../../../../../components/form/FormField";
 import { FieldBreadcrumbs } from "./FieldBreadcrumbs";
 import { FieldsDropdown } from "./FieldsDropdown";
 import { InvertedIndexAnalyzerDropdown } from "./InvertedIndexAnalyzerDropdown";
