+++ js/apps/system/_admin/aardvark/APP/react/src/views/collections/indices/addIndex/invertedIndex/InvertedIndexLeftPane.tsx
@@ -7,7 +7,7 @@ import {
   Box
 } from "@chakra-ui/react";
 import React from "react";
-import { FormFieldProps } from "../FormField";
+import { FormFieldProps } from "../../../../../components/form/FormField";
 import { IndexFormFieldsList } from "../IndexFormFieldList";
 import { InvertedIndexAnalyzerDropdown } from "./InvertedIndexAnalyzerDropdown";
 import { InvertedIndexConsolidationPolicy } from "./InvertedIndexConsolidationPolicy";
