+++ js/apps/system/_admin/aardvark/APP/react/src/views/collections/indices/addIndex/invertedIndex/InvertedIndexStoredValues.tsx
@@ -4,7 +4,7 @@ import { FieldArray, useFormikContext } from "formik";
 import React from "react";
 import { CreatableMultiSelectControl } from "../../../../../components/form/CreatableMultiSelectControl";
 import { SelectControl } from "../../../../../components/form/SelectControl";
-import { FormFieldProps } from "../FormField";
+import { FormFieldProps } from "../../../../../components/form/FormField";
 import { InvertedIndexValuesType } from "./useCreateInvertedIndex";
 
 export const InvertedIndexStoredValues = ({
